requirements.txt
multiprocess~=0.70.14
discord~=2.1.0
RestrictedPython~=6.0
pathos~=0.3.0
python-dotenv~=0.21.0